test/web-platform-tests/to-upstream/css/css-borders/border-color.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>border color</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-shorthands">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/cssstyle/issues/251 -->

<script>
"use strict";

test(() => {
  const div = document.createElement("div");
  div.style.border = "1px solid #ffffff";
  assert_equals(div.style.length > 0, true, "length");
  assert_equals(div.style.border, "1px solid rgb(255, 255, 255)", "border");
  assert_equals(div.style.borderWidth, "1px", "border-width");
  assert_equals(div.style.borderStyle, "solid", "border-style");
  assert_equals(div.style.borderColor, "rgb(255, 255, 255)", "border-color");
  assert_equals(div.style.borderImage, "none", "border-image");
  assert_equals(div.style.cssText, "border: 1px solid rgb(255, 255, 255);", "cssText");
  assert_equals(div.outerHTML, '<div style="border: 1px solid rgb(255, 255, 255);"></div>', "outerHTML");
}, "Set border as 1px solid #ffffff");
</script>
